backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
black==23.11.0
flake8==6.1.0
ruff==0.1.6
redis==5.0.1
aiohttp==3.9.1
stellar-sdk==13.1.0
argon2-cffi==23.1.0
fastapi-mail==1.4.1
Pillow==10.4.0
